+++ lib/phoenix_live_view/test/structs.ex
@@ -7,11 +7,13 @@ defmodule Phoenix.LiveViewTest.View do
     * `id` - The DOM id of the LiveView
     * `module` - The module of the running LiveView
     * `pid` - The Pid of the running LiveView
+    * `endpoint` - The endpoint for the LiveView
 
   See the `Phoenix.LiveViewTest` documentation for usage.
   """
   defstruct id: nil,
             module: nil,
             pid: nil,
-            proxy: nil
+            proxy: nil,
+            endpoint: nil
 end